+++ crates/nu-engine/src/filesystem/filesystem_shell.rs
@@ -3,8 +3,8 @@ use crate::filesystem::dir_info::{DirBuilder, DirInfo};
 use crate::filesystem::path::canonicalize;
 use crate::filesystem::utils::FileStructure;
 use crate::maybe_text_codec::{MaybeTextCodec, StringOrBinary};
-use crate::shell::shell::Shell;
 use crate::shell::shell_args::{CdArgs, CopyArgs, LsArgs, MkdirArgs, MvArgs, RemoveArgs};
+use crate::shell::Shell;
 use encoding_rs::Encoding;
 use futures::stream::BoxStream;
 use futures::StreamExt;
