+++ crates/nu-engine/src/plugin/mod.rs
@@ -1,2 +1,2 @@
-pub mod plugin;
+pub mod build_plugin;
 pub(crate) mod run_plugin;
