+++ crates/nu-engine/src/shell/help_shell.rs
@@ -1,8 +1,8 @@
 use crate::command_args::EvaluatedWholeStreamCommandArgs;
 use crate::evaluate::scope::Scope;
 use crate::maybe_text_codec::StringOrBinary;
-use crate::shell::shell::Shell;
 use crate::shell::shell_args::{CdArgs, CopyArgs, LsArgs, MkdirArgs, MvArgs, RemoveArgs};
+use crate::shell::Shell;
 use crate::whole_stream_command::Command;
 use encoding_rs::Encoding;
 use futures::stream::BoxStream;
