+++ crates/nu-engine/src/shell/shell_manager.rs
@@ -1,6 +1,6 @@
 use crate::command_args::EvaluatedWholeStreamCommandArgs;
 use crate::maybe_text_codec::StringOrBinary;
-use crate::shell::shell::Shell;
+use crate::shell::Shell;
 use futures::Stream;
 use nu_stream::OutputStream;
